requirements.txt
# YT-Short-Clipper Dependencies
# ==============================

# OpenAI API (supports both OpenAI and Azure OpenAI)
openai>=1.0.0

# Environment variables
python-dotenv>=1.0.0

# Computer Vision (face detection, speaker tracking)
opencv-python>=4.8.0

# Numerical operations
numpy>=1.24.0

# Speech recognition (auto captions)
openai-whisper>=20231117

# Note: System dependencies required (not installable via pip):
# - FFmpeg 4.4+ (video processing)
# - yt-dlp (YouTube downloading) - can also install via: pip install yt-dlp
